Feat(specialties): New component created and separate data arrays in appropriate file
