+++ src/styles/GlobalStyle.js
@@ -55,6 +55,7 @@ const GlobalStyle = createGlobalStyle`
     }
     a {
         text-decoration: none;
+        color: inherit;
     }
 `;
